Fix NPE with new/empty threadfix.properties file
Fix NullPointerException in ZapPropertiesManager with a new/empty
threadfix.properties file (as when starting to use the add-on), the
variables autoSpider and useHttps were being accessed even if null.
Change to use Boolean.parseBoolean to parse the variables, which uses
the required logic (checks if not null and if true, ignoring case).